+++ CHANGELOG.md
@@ -352,3 +352,4 @@
 - 2024-03-30: Rate README, follow up, mutate, sender, +3
 - 2024-04-01: Docs summaries, site config, mod, send, +3
 - 2024-04-02: Init mod, site config, mod, summaries, +3
+- 2024-04-03: Add leads, mod, site config, mailbox, +6 db